# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoxForge API",
    "description": "Text-to-speech synthesis API with voice profile management",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/chapters": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Chapters",
        "operationId": "list_chapters_api_projects__project_id__chapters_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChapterResponse"
                  },
                  "title": "Response List Chapters Api Projects  Project Id  Chapters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Chapter",
        "operationId": "create_chapter_api_projects__project_id__chapters_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/chapters/{chapter_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Chapter",
        "operationId": "update_chapter_api_projects_chapters__chapter_id__patch",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chapter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Chapter",
        "operationId": "delete_chapter_api_projects_chapters__chapter_id__delete",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/{chapter_id}/synthesize": {
      "post": {
        "tags": [
          "chapter-synth"
        ],
        "summary": "Synthesize a chapter with per-chunk tracking",
        "operationId": "synthesize_chapter_api_chapters__chapter_id__synthesize_post",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{chapter_id}/regenerate-chunk/{chunk_index}": {
      "post": {
        "tags": [
          "chapter-synth"
        ],
        "summary": "Regenerate a single chunk of a chapter",
        "description": "Regenerate chunk N without re-synthesizing the whole chapter.\n\nFinds the latest generation for this chapter, re-splits the text,\nsynthesizes only the requested chunk, replaces the take, and\nre-exports the full audio by splicing in the new chunk.",
        "operationId": "regenerate_chunk_api_chapters__chapter_id__regenerate_chunk__chunk_index__post",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chapter Id"
            }
          },
          {
            "name": "chunk_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chunk Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{chapter_id}/chunks": {
      "get": {
        "tags": [
          "chapter-synth"
        ],
        "summary": "Get chunk map for latest generation",
        "description": "Return the chunk list for the latest generation of a chapter.",
        "operationId": "get_chunk_map_api_chapters__chapter_id__chunks_get",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkMapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{chapter_id}/upload-audio": {
      "post": {
        "tags": [
          "chapter-synth"
        ],
        "summary": "Upload a pre-recorded audio file as a generation of this chapter",
        "description": "Save the uploaded audio under ``data/output/`` and register a\n``generations`` row with ``engine=\"upload\"`` and ``status=\"done\"``.\n\nMakes the chapter behave identically to a TTS-synthesized one\ndownstream: Studio can load it, Workbench can render a video from\nit, export can bundle it. Browser recordings (webm/opus) and other\ncontainers are accepted — ``AudioSegment.from_file`` handles them\nvia ffmpeg. The file is stored as-is, without transcoding, to\npreserve quality.",
        "operationId": "upload_chapter_audio_api_chapters__chapter_id__upload_audio_post",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chapter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_chapter_audio_api_chapters__chapter_id__upload_audio_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedChapterGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_chapter_audio_api_chapters__chapter_id__upload_audio_post": {
        "properties": {
          "audio": {
            "type": "string",
            "format": "binary",
            "title": "Audio"
          }
        },
        "type": "object",
        "required": [
          "audio"
        ],
        "title": "Body_upload_chapter_audio_api_chapters__chapter_id__upload_audio_post"
      },
...
      "ChapterCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": "Chapter"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          }
        },
        "type": "object",
        "title": "ChapterCreate"
      },
      "ChapterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "active_generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Generation Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "text",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "title": "ChapterResponse"
      },
      "ChapterUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "active_generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Generation Id"
          }
        },
        "type": "object",
        "title": "ChapterUpdate"
      },
...
      "ChunkInfo": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "take_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Id"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "index",
          "text",
          "status",
          "duration"
        ],
        "title": "ChunkInfo"
      },
      "ChunkMapResponse": {
        "properties": {
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkInfo"
            },
            "type": "array",
            "title": "Chunks"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "chunks",
          "total"
        ],
        "title": "ChunkMapResponse"
      },
...
      "DeletedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "id"
        ],
        "title": "DeletedResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadedChapterGenerationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "chapter_id": {
            "type": "string",
            "title": "Chapter Id"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "output_format": {
            "type": "string",
            "title": "Output Format"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chapter_id",
          "engine",
          "status",
          "duration",
          "file_path",
          "output_format"
        ],
        "title": "UploadedChapterGenerationResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}